Replaced old ETL impl with new impl
