+++ src/connection/etl/export/reader.rs
@@ -1,179 +1,262 @@
 use std::{
-    fmt::Debug,
-    io::{Read, Result as IoResult},
-    pin::Pin,
+    future::Future,
+    io::{Error as IoError, ErrorKind as IoErrorKind, Result as IoResult},
+    pin::{pin, Pin},
+    sync::Mutex,
     task::{ready, Context, Poll},
 };
 
-use futures_io::{AsyncBufRead, AsyncRead, AsyncWrite, IoSliceMut};
-use futures_util::io::BufReader;
+use futures_channel::mpsc::{Receiver, SendError, Sender};
+use futures_io::{AsyncBufRead, AsyncRead};
+use futures_util::{
+    stream::{Forward, IntoAsyncRead},
+    FutureExt, Sink, SinkExt, Stream, StreamExt, TryStreamExt,
+};
+use http_body_util::Empty;
+use hyper::{
+    body::{Body, Bytes, Incoming},
+    header::CONNECTION,
+    server::conn::http1::{Builder, Connection},
+    service::Service,
+    Request, Response, StatusCode,
+};
 use pin_project::pin_project;
 
-use crate::{
-    connection::websocket::socket::ExaSocket,
-    etl::{error::ExaEtlError, traits::EtlBufReader},
-};
+use crate::connection::websocket::socket::ExaSocket;
 
-/// Low-level async reader used to read chunked HTTP data from Exasol.
+type ExportResponse = Response<Empty<&'static [u8]>>;
+pub type ExportConnection = Connection<ExaSocket, ExportService>;
+pub type ExportReader = IntoAsyncRead<ReaderStream>;
+
 #[pin_project]
-#[derive(Debug)]
-pub struct ExportReader {
-    #[pin]
-    socket: BufReader<ExaSocket>,
-    state: ReaderState,
-    chunk_size: usize,
-}
+struct ExportStream(#[pin] Incoming);
 
-impl ExportReader {
-    /// HTTP Response for the EXPORT request Exasol sends.
-    const RESPONSE: &'static [u8; 38] = b"HTTP/1.1 200 OK\r\n\
-                                  Connection: close\r\n\
-                                  \r\n";
+impl Stream for ExportStream {
+    type Item = IoResult<Bytes>;
 
-    pub fn new(socket: ExaSocket, buffer_size: usize) -> Self {
-        Self {
-            socket: BufReader::with_capacity(buffer_size, socket),
-            state: ReaderState::SkipRequest([0; 4]),
-            chunk_size: 0,
+    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+        match ready!(self.project().0.poll_frame(cx)) {
+            Some(Ok(f)) => Poll::Ready(f.into_data().ok().map(Ok)),
+            Some(Err(e)) => Poll::Ready(Some(Err(map_hyper_err(e)))),
+            None => Poll::Ready(None),
         }
     }
 }
 
-impl AsyncRead for ExportReader {
+pub struct ExportSink(Sender<Bytes>);
+
+impl Sink<Bytes> for ExportSink {
+    type Error = IoError;
+
+    fn poll_ready(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.0.poll_ready_unpin(cx).map_err(map_send_error)
+    }
+
+    fn start_send(mut self: Pin<&mut Self>, item: Bytes) -> Result<(), Self::Error> {
+        self.0.start_send_unpin(item).map_err(map_send_error)
+    }
+
+    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.0.poll_flush_unpin(cx).map_err(map_send_error)
+    }
+
+    fn poll_close(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.0.poll_close_unpin(cx).map_err(map_send_error)
+    }
+}
+
+pub struct ExportFuture(Forward<ExportStream, ExportSink>);
+
+impl Future for ExportFuture {
+    type Output = IoResult<ExportResponse>;
+
+    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        match ready!(self.0.poll_unpin(cx)) {
+            Ok(()) => (),
+            Err(e) => return Poll::Ready(Err(e)),
+        };
+
+        let response = Response::builder()
+            .header(CONNECTION, "close")
+            .status(StatusCode::OK)
+            .body(Empty::new())
+            .map_err(map_http_error);
+
+        Poll::Ready(response)
+    }
+}
+
+pub struct ExportService(Mutex<Option<ExportSink>>);
+
+impl ExportService {
+    pub fn new(sink: ExportSink) -> Self {
+        Self(Mutex::new(Some(sink)))
+    }
+}
+
+impl Service<Request<Incoming>> for ExportService {
+    type Response = ExportResponse;
+
+    type Error = IoError;
+
+    type Future = ExportFuture;
+
+    fn call(&self, req: Request<Incoming>) -> Self::Future {
+        let sink = self.0.lock().unwrap().take().unwrap();
+        ExportFuture(ExportStream(req.into_body()).forward(sink))
+    }
+}
+
+#[derive(Debug)]
+pub struct ReaderStream(Receiver<Bytes>);
+
+impl Stream for ReaderStream {
+    type Item = IoResult<Bytes>;
+
+    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+        match ready!(self.0.poll_next_unpin(cx)) {
+            Some(b) => Poll::Ready(Some(Ok(b))),
+            None => Poll::Ready(None),
+        }
+    }
+}
+
+#[derive(Debug)]
+#[pin_project]
+pub struct ExaReader {
+    #[pin]
+    reader: ExportReader,
+    conn: ExportConnection,
+    state: ExaReaderState,
+}
+
+impl ExaReader {
+    pub fn new(socket: ExaSocket) -> Self {
+        let (data_tx, data_rx): (_, Receiver<Bytes>) = futures_channel::mpsc::channel(0);
+        let service = ExportService::new(ExportSink(data_tx));
+        let reader = ReaderStream(data_rx).into_async_read();
+        let conn = Builder::new().serve_connection(socket, service);
+
+        Self {
+            reader,
+            conn,
+            state: ExaReaderState::Reading,
+        }
+    }
+
+    fn poll_conn(&mut self, cx: &mut Context<'_>) -> Poll<IoResult<usize>> {
+        let res = match ready!(self.conn.poll_unpin(cx)) {
+            Ok(()) => Ok(0),
+            Err(err) => Err(map_hyper_err(err)),
+        };
+
+        self.state = ExaReaderState::Done;
+        Poll::Ready(res)
+    }
+
+    fn poll_read_internal(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: &mut [u8],
+    ) -> Poll<IoResult<usize>> {
+        let _ = self.as_mut().poll_conn(cx)?;
+        let mut this = self.as_mut().project();
+
+        match ready!(this.reader.as_mut().poll_read(cx, buf)) {
+            Ok(0) => {
+                *this.state = ExaReaderState::Responding;
+                self.poll_conn(cx)
+            }
+            Ok(n) => Poll::Ready(Ok(n)),
+            Err(e) => {
+                *this.state = ExaReaderState::Done;
+                Poll::Ready(Err(e))
+            }
+        }
+    }
+
+    fn poll_read_vectored_internal(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        bufs: &mut [std::io::IoSliceMut<'_>],
+    ) -> Poll<IoResult<usize>> {
+        let _ = self.as_mut().poll_conn(cx)?;
+        let mut this = self.as_mut().project();
+
+        match ready!(this.reader.as_mut().poll_read_vectored(cx, bufs)) {
+            Ok(0) => {
+                *this.state = ExaReaderState::Responding;
+                self.poll_conn(cx)
+            }
+            Ok(n) => Poll::Ready(Ok(n)),
+            Err(e) => {
+                *this.state = ExaReaderState::Done;
+                Poll::Ready(Err(e))
+            }
+        }
+    }
+}
+
+#[derive(Debug)]
+enum ExaReaderState {
+    Reading,
+    Responding,
+    Done,
+}
+
+impl AsyncRead for ExaReader {
     fn poll_read(
         mut self: Pin<&mut Self>,
         cx: &mut Context<'_>,
         buf: &mut [u8],
     ) -> Poll<IoResult<usize>> {
-        let mut rem = ready!(self.as_mut().poll_fill_buf(cx))?;
-        let nread = rem.read(buf)?;
-        self.consume(nread);
-
-        Poll::Ready(Ok(nread))
+        match self.state {
+            ExaReaderState::Reading => self.poll_read_internal(cx, buf),
+            ExaReaderState::Responding => self.poll_conn(cx),
+            ExaReaderState::Done => Poll::Ready(Ok(0)),
+        }
     }
 
     fn poll_read_vectored(
         mut self: Pin<&mut Self>,
         cx: &mut Context<'_>,
-        bufs: &mut [IoSliceMut<'_>],
+        bufs: &mut [std::io::IoSliceMut<'_>],
     ) -> Poll<IoResult<usize>> {
-        let mut rem = ready!(self.as_mut().poll_fill_buf(cx))?;
-        let nread = rem.read_vectored(bufs)?;
-        self.consume(nread);
-
-        Poll::Ready(Ok(nread))
+        match self.state {
+            ExaReaderState::Reading => self.poll_read_vectored_internal(cx, bufs),
+            ExaReaderState::Responding => self.poll_conn(cx),
+            ExaReaderState::Done => Poll::Ready(Ok(0)),
+        }
     }
 }
 
-impl AsyncBufRead for ExportReader {
+impl AsyncBufRead for ExaReader {
     fn poll_fill_buf(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<IoResult<&[u8]>> {
-        loop {
-            let mut this = self.as_mut().project();
-
-            match this.state {
-                ReaderState::SkipRequest(buf) => {
-                    ready!(this.socket.poll_until_double_crlf(cx, buf))?;
-                    *this.state = ReaderState::ReadSize;
-                }
-
-                ReaderState::ReadSize => {
-                    let byte = ready!(this.socket.poll_read_byte(cx))?;
-
-                    let digit = match byte {
-                        b'0'..=b'9' => byte - b'0',
-                        b'a'..=b'f' => 10 + byte - b'a',
-                        b'A'..=b'F' => 10 + byte - b'A',
-                        b'\r' => {
-                            *this.state = ReaderState::ExpectSizeLF;
-                            continue;
-                        }
-                        _ => Err(ExaEtlError::InvalidChunkSizeByte(byte))?,
-                    };
-
-                    *this.chunk_size = this
-                        .chunk_size
-                        .checked_mul(16)
-                        .and_then(|size| size.checked_add(digit.into()))
-                        .ok_or(ExaEtlError::ChunkSizeOverflow)?;
-                }
-
-                ReaderState::ReadData => {
-                    if *this.chunk_size > 0 {
-                        let this = self.get_mut();
-                        let buf = ready!(Pin::new(&mut this.socket).poll_fill_buf(cx))?;
-                        let max_read = buf.len().min(this.chunk_size);
-                        return Poll::Ready(Ok(&buf[..max_read]));
-                    }
-
-                    *this.state = ReaderState::ExpectDataCR;
-                }
-
-                ReaderState::ExpectSizeLF => {
-                    ready!(this.socket.poll_read_lf(cx))?;
-                    // We just read the chunk size separator.
-                    // If the chunk size is 0 at this stage, then
-                    // this is the final, empty, chunk so we'll end the reader.
-                    if *this.chunk_size == 0 {
-                        *this.state = ReaderState::ExpectEndCR;
-                    } else {
-                        *this.state = ReaderState::ReadData;
-                    }
-                }
-
-                ReaderState::ExpectDataCR => {
-                    ready!(this.socket.poll_read_cr(cx))?;
-                    *this.state = ReaderState::ExpectDataLF;
-                }
-
-                ReaderState::ExpectDataLF => {
-                    ready!(this.socket.poll_read_lf(cx))?;
-                    *this.state = ReaderState::ReadSize;
-                }
-
-                ReaderState::ExpectEndCR => {
-                    ready!(this.socket.poll_read_cr(cx))?;
-                    *this.state = ReaderState::ExpectEndLF;
-                }
-
-                ReaderState::ExpectEndLF => {
-                    ready!(this.socket.poll_read_lf(cx))?;
-                    *this.state = ReaderState::WriteResponse(0);
-                }
-
-                ReaderState::WriteResponse(offset) => {
-                    // EOF is reached after writing the HTTP response.
-                    let socket = this.socket.as_mut();
-                    ready!(socket.poll_send_static(cx, Self::RESPONSE, offset))?;
-                    *this.state = ReaderState::End;
-                }
-
-                ReaderState::End => {
-                    // We flush to ensure that all data has reached Exasol
-                    // and the reader can finish or even be dropped.
-                    ready!(this.socket.poll_flush(cx))?;
-                    return Poll::Ready(Ok(&[]));
-                }
-            };
-        }
+        let _ = self.as_mut().poll_conn(cx)?;
+        self.project().reader.poll_fill_buf(cx)
     }
 
     fn consume(self: Pin<&mut Self>, amt: usize) {
-        let this = self.project();
-        this.socket.consume(amt);
-        *this.chunk_size -= amt;
+        self.project().reader.consume(amt)
     }
 }
 
-#[derive(Copy, Clone, Debug)]
-enum ReaderState {
-    SkipRequest([u8; 4]),
-    ReadSize,
-    ReadData,
-    ExpectSizeLF,
-    ExpectDataCR,
-    ExpectDataLF,
-    ExpectEndCR,
-    ExpectEndLF,
-    WriteResponse(usize),
-    End,
+fn map_hyper_err(err: hyper::Error) -> IoError {
+    let kind = if err.is_timeout() {
+        IoErrorKind::TimedOut
+    } else if err.is_parse_too_large() {
+        IoErrorKind::InvalidData
+    } else {
+        IoErrorKind::BrokenPipe
+    };
+
+    IoError::new(kind, err)
+}
+
+fn map_send_error(err: SendError) -> IoError {
+    IoError::new(IoErrorKind::BrokenPipe, err)
+}
+
+fn map_http_error(err: hyper::http::Error) -> IoError {
+    IoError::new(IoErrorKind::BrokenPipe, err)
 }
