+++ src/connection/etl/mod.rs
@@ -90,6 +90,7 @@ use arrayvec::ArrayString;
 pub use export::{ExaExport, ExportBuilder, ExportSource};
 use futures_core::future::BoxFuture;
 use futures_io::{AsyncRead, AsyncWrite};
+use futures_util::future::{select, try_join_all, Either};
 use hyper::rt;
 pub use import::{ExaImport, ImportBuilder, Trim};
 pub use row_separator::RowSeparator;
@@ -144,9 +145,6 @@ where
     let cmd = ExaCommand::new_execute(&query, &con.ws.attributes).try_into()?;
     con.ws.send(cmd).await?;
 
-    // Create the ETL workers
-    let sockets = job.create_workers(futures, with_compression);
-
     // Query execution driving future to be returned and awaited
     // alongside the worker IO operations
     let future = Box::pin(async move {
@@ -156,6 +154,14 @@ where
             QueryResult::RowCount { row_count } => Ok(ExaQueryResult::new(row_count)),
         }
     });
+
+    let (sockets, future): (_, JobFuture<'_>) = match select(future, try_join_all(futures)).await {
+        Either::Left((res, _)) => (Vec::new(), Box::pin(async move { res })),
+        Either::Right((sockets, future)) => (sockets?, future),
+    };
+
+    // Create the ETL workers
+    let sockets = job.create_workers(sockets, with_compression);
 
     Ok((future, sockets))
 }
