+++ src/connection/etl/traits/mod.rs
@@ -1,7 +1,5 @@
-mod buf_reader;
 mod job;
 mod with_socket_maker;
 
-pub use buf_reader::EtlBufReader;
 pub use job::EtlJob;
 pub use with_socket_maker::WithSocketMaker;
